+++ spec/lib/memcached-store-spec.js
@@ -62,6 +62,17 @@ describe('set', function () {
       })
     })
   })
+
+  it('should store a value & return promise if no callback provided', function (done) {
+    var result = memcachedCache.set('foo', 'bar')
+    expect(result.then).toBeInstanceOf(Function)
+    result.then(function (ok) {
+      expect(ok).toBe(true)
+      done()
+    }).catch(function (err) {
+      done(err)
+    })
+  })
 })
 
 describe('get', function () {
@@ -86,6 +97,21 @@ describe('get', function () {
       })
     })
   })
+
+  it('should retrieve a value for a given key & return promise if no cb passed', function (done) {
+    var value = 'bar'
+    memcachedCache.set('foo', value)
+      .then(function () {
+        var res = memcachedCache.get('foo')
+        expect(res.then).toBeInstanceOf(Function)
+        return res
+      }).then(function (result) {
+        expect(result).toBe(value)
+        done()
+      }).catch(function (err) {
+        done(err)
+      })
+  })
 })
 
 describe('del', function () {
@@ -98,11 +124,19 @@ describe('del', function () {
     })
   })
 
-  it('should delete a value for a given key without callback', function (done) {
-    memcachedCache.set('foo', 'bar', function () {
-      memcachedCache.del('foo')
-      done()
-    })
+  it('should delete a value for a given key & return promise if no cb passed', function (done) {
+    var value = 'bar'
+    memcachedCache.set('foo', value)
+      .then(function () {
+        var res = memcachedCache.del('foo')
+        expect(res.then).toBeInstanceOf(Function)
+        return res
+      }).then(function (result) {
+        expect(result).toBe(null)
+        done()
+      }).catch(function (err) {
+        done(err)
+      })
   })
 })
 
@@ -119,6 +153,23 @@ describe('reset', function () {
         })
       })
     })
+  })
+
+  it('should flush underlying db & return promise if no cb passed', function (done) {
+    memcachedCache.set('foo', 'bar')
+      .then(function () {
+        var res = memcachedCache.reset()
+        expect(res.then).toBeInstanceOf(Function)
+        return res
+      }).then(function (result) {
+        expect(result).toBe(null)
+        return memcachedCache.get('foo')
+      }).then(function (result) {
+        expect(result).toBe(null)
+        done()
+      }).catch(function (err) {
+        done(err)
+      })
   })
 })